Implemented sizers to allow for resizing of the window during runtime

+++ main.cpp
@@ -1,10 +1,8 @@
-// Start of wxWidgets "Hello World" Program
 #include <iostream>
 #include  <wx/wx.h>
 #include <wx/artprov.h>
 #include "Database.h"
 #include "Task.h"
-#include <cstddef>
 
 using namespace std;
 
@@ -27,7 +25,8 @@ public:
 private:
     int TASK_Y = -50;
     bool CAN_EDIT = true;
-    wxBoxSizer* mainSizer = new wxBoxSizer(wxVERTICAL);;
+    wxBoxSizer* mainSizer = new wxBoxSizer(wxVERTICAL);
+    wxBoxSizer* taskContainer = new wxBoxSizer(wxVERTICAL);
     wxButton* addTask;
     wxTextCtrl* typeTask;
     wxBitmapButton* editTaskButton;
@@ -39,19 +38,19 @@ private:
     void OnAddTask(wxCommandEvent const& event);
     void OnEnterNew(wxCommandEvent const& event);
     void OnEnterEdit(wxCommandEvent const& event);
-    void OnEditTask(wxCommandEvent const& event);
+    void OnEditButtonClick(wxCommandEvent const& event);
     void OnRemoveTask(wxCommandEvent const& event);
     void CreateTask();
-    void PopulateTasks();
+    void LoadTasks();
     void SetupSizer();
 };
 
-
-//"this function is called  upon application startup and creates the main window"
 bool App::OnInit()
 {
     //"there is no memory leak in this: wxWidgets takes ownership of all the window objects and destroys them automatically when the corresponding on-screen window is destroyed."
+    // Using new here is fine since wxWidgets automatically manages memory
     auto *frame = new MainFrame();
+    // TODO - add these sizes as constants
     frame->SetClientSize(1000, 600);
      //"frames are hidden by default to allow filling them to be filled with contents before showing everything at once."
     //"Show() must be called for the application to appear"
@@ -73,72 +72,39 @@ bool App::OnInit()
     menuBar->Append(menuFile, "&File");
     menuBar->Append(menuHelp, "&Help");
 
-    PopulateTasks();
+     LoadTasks();
 
-    addTask = new wxButton(this, wxID_ADD, "Add Task", wxPoint(50,500), wxSize(900,50));
+    addTask = new wxButton(this, wxID_ADD, "Add Task", wxDefaultPosition, wxSize(900,50));
     addTask->SetBitmap(wxArtProvider::GetBitmap(wxART_PLUS, wxART_BUTTON));
     addTask->Bind(wxEVT_BUTTON, &MainFrame::OnAddTask, this, wxID_ADD);
 
+    SetupSizer();
     //"We have to connect our event handlers to the events we want to handle in them. We do this by calling Bind() to send all the menu events (identified by wxEVT_MENU event type) with the specified ID to the given function."
     Bind(wxEVT_MENU, &MainFrame::OnExit, this, wxID_EXIT);
 
 }
 
-void MainFrame::SetupSizer() {
-}
-
-void MainFrame::PopulateTasks() {
+void MainFrame::LoadTasks() {
     int count = DB->CountEntries();
     vector<string> descriptions = DB->GetDescription();
     for (int i = 0; i < count; i++) {
-        TASK_Y += 50;
-        task = new Task(this, TASK_Y, descriptions[i]);
+        TASK_Y += 50; // TODO - should be a const; If refactor doesn't autosize task, then create taskSize consts
+        task = new Task(this, TASK_Y, descriptions[i], taskContainer);
 
         tasks.push_back(task);
         editTaskButton = task->GetEditButton();
         removeTaskButton = task->GetRemoveButton();
-        editTaskButton->Bind(wxEVT_BUTTON, &MainFrame::OnEditTask, this, wxID_EDIT);
+        editTaskButton->Bind(wxEVT_BUTTON, &MainFrame::OnEditButtonClick, this, wxID_EDIT);
         removeTaskButton->Bind(wxEVT_BUTTON, &MainFrame::OnRemoveTask, this, wxID_REMOVE);
     }
 }
 
-void MainFrame::CreateTask() {
-    task = new Task(this, TASK_Y);
-    tasks.push_back(task);
-
-    typeTask = task->GetTextCtrl();
-    typeTask->Bind(wxEVT_TEXT_ENTER, &MainFrame::OnEnterNew, this);
-    removeTaskButton = task->GetRemoveButton();
-    removeTaskButton->Bind(wxEVT_BUTTON, &MainFrame::OnRemoveTask, this, wxID_REMOVE);
-}
-
-void MainFrame::OnEnterNew(wxCommandEvent const &event) {
-    wxString taskDescription = typeTask->GetValue();
-    DB->InsertData(taskDescription);
-
-    editTaskButton = task->PublishTask(taskDescription);
-    editTaskButton->Bind(wxEVT_BUTTON, &MainFrame::OnEditTask, this, wxID_EDIT);
-
-    CAN_EDIT = true;
-}
-
-void MainFrame::OnEnterEdit(wxCommandEvent const &event) {
-    wxString taskDescription = typeTask->GetValue();
-    int index = (typeTask->GetParent()->GetPosition().y) / 50;
-    DB->EditData(index, taskDescription);
-
-    editTaskButton = tasks[index]->PublishEditedTask(taskDescription);
-    editTaskButton->Bind(wxEVT_BUTTON, &MainFrame::OnEditTask, this, wxID_EDIT);
-
-    CAN_EDIT = true;
-}
-
-void MainFrame::OnExit(wxCommandEvent const& event) {
-    for (int i = 0; i < tasks.size(); i++) {
-        delete tasks[i];
-    }
-    delete DB;
-    Close(true);
+void MainFrame::SetupSizer() {
+    mainSizer->Add(taskContainer, wxSizerFlags().Expand());
+    mainSizer->AddStretchSpacer();
+    mainSizer->Add(addTask, wxSizerFlags().Expand().Border(wxALL, 25));
+    SetSizer(mainSizer);
+    mainSizer->SetSizeHints(this);
 }
 
 void MainFrame::OnAddTask(wxCommandEvent const& event) {
@@ -149,33 +115,86 @@ void MainFrame::OnAddTask(wxCommandEvent const& event) {
     }
 }
 
-void MainFrame::OnEditTask(wxCommandEvent const& event) {
+void MainFrame::CreateTask() {
+    task = new Task(this, TASK_Y, taskContainer);
+    tasks.push_back(task);
+
+    typeTask = task->GetTextCtrl(); // TODO - rename to something better like textBox
+    removeTaskButton = task->GetRemoveButton();
+    typeTask->Bind(wxEVT_TEXT_ENTER, &MainFrame::OnEnterNew, this); // TODO - what is winid?
+    removeTaskButton->Bind(wxEVT_BUTTON, &MainFrame::OnRemoveTask, this, wxID_REMOVE);
+
+    Layout();
+}
+// TODO - consider renaming to OnKeyboardEnterNew
+void MainFrame::OnEnterNew(wxCommandEvent const &event) { // TODO - figure out if you need this param, if not then comment why unsused and needed
+    wxString taskDescription = typeTask->GetValue();
+    DB->InsertData(taskDescription);
+
+    editTaskButton = task->PublishTask(taskDescription);
+    editTaskButton->Bind(wxEVT_BUTTON, &MainFrame::OnEditButtonClick, this, wxID_EDIT);
+
+    CAN_EDIT = true;
+    Layout();
+}
+
+// TODO - consider renaming to OnEditButtonClick()
+void MainFrame::OnEditButtonClick(wxCommandEvent const& event) {
     if (CAN_EDIT) {
         CAN_EDIT = false;
         wxObject const* obj = event.GetEventObject();
         auto const* button = wxDynamicCast(obj, wxBitmapButton);
-        int index = (button->GetParent()->GetPosition().y) / 50;
+        auto const* currentPanel = button->GetParent();
+        int index = (currentPanel->GetPosition().y) / currentPanel->GetSize().y;
         wxString label = tasks[index]->GetCheckBox()->GetLabel();
 
         typeTask = tasks[index]->EditTask(label);
         typeTask->Bind(wxEVT_TEXT_ENTER, &MainFrame::OnEnterEdit, this);
+        Layout();
     }
+}
+// TODO - consider renaming to OnKeyboardEnterEditTask / onEditTask
+// TODO - figure out if this only can happen on enter. What happens on Tab or mouse/cursor leave?
+void MainFrame::OnEnterEdit(wxCommandEvent const &event) {
+    wxString taskDescription = typeTask->GetValue();
+    auto const* currentPanel = typeTask->GetParent();
+    int index = (currentPanel->GetPosition().y) / currentPanel->GetSize().y;
+    DB->EditData(index, taskDescription);
+
+    editTaskButton = tasks[index]->PublishTask(taskDescription);
+    editTaskButton->Bind(wxEVT_BUTTON, &MainFrame::OnEditButtonClick, this, wxID_EDIT);
+
+    CAN_EDIT = true;
+    Layout();
 }
 
 void MainFrame::OnRemoveTask(wxCommandEvent const& event) {
     const wxObject* obj = event.GetEventObject();
     auto const* button = wxDynamicCast(obj, wxBitmapButton);
-    int index = (button->GetParent()->GetPosition().y) / 50;
-    //delete pre-established taskCheckBoxes
-    tasks[index]->DestroyPanel();
-    delete tasks[index];
-    tasks.erase(tasks.begin() + index);
-    DB->RemoveData(index);
+    auto const* currentPanel = button->GetParent();
+    if (CAN_EDIT || typeTask->GetParent() == currentPanel) {
+        int index = (currentPanel->GetPosition().y) / currentPanel->GetSize().y;
+        //delete pre-established taskCheckBoxes
+        tasks[index]->DestroyPanel();
+        delete tasks[index];
+        tasks.erase(tasks.begin() + index);
+        DB->RemoveData(index);
 
-    for (int i = index ; i < tasks.size(); i++) {
-        tasks[i]->MovePanel();
+        for (int i = index ; i < tasks.size(); i++) {
+            tasks[i]->MovePanel();
+        }
+
+        TASK_Y -= 50;
+        CAN_EDIT = true;
+        Layout();
     }
-
-    TASK_Y -= 50;
-    CAN_EDIT = true;
 }
+// OnAppClose
+void MainFrame::OnExit(wxCommandEvent const& event) {
+    for (Task* current : tasks) {
+        delete current;
+    }
+    delete DB;
+    Close(true);
+}
+
